% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_format.r, R/change_format_backup.r,
%   R/change_format_wrong.r
\name{change_format}
\alias{change_format}
\title{Function to make the format readable to WC portal.}
\usage{
change_format(data)

change_format(data)

change_format(data)
}
\description{
Function to make the format readable to WC portal.

Function to make the format readable to WC portal.

Function to make the format readable to WC portal.
}
\examples{
kineisdata=download_kineis_data(device_refs='267094',from_date="2026-02-01",to_date="2026-02-05")
head(kineisdata)
kineisdata_formatted=change_format(kineisdata)
kineisdata=download_kineis_data(device_refs='267094',from_date="2026-02-01",to_date="2026-02-05")
head(kineisdata)
kineisdata_formatted=change_format(kineisdata)
kineisdata=download_kineis_data(device_refs='267094',from_date="2026-02-01",to_date="2026-02-05")
head(kineisdata)
kineisdata_formatted=change_format(kineisdata)
}
