% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_credentials.r
\name{check_credentials}
\alias{check_credentials}
\title{Function to check is credentials are stored in a file.
It also provides the access token
If not, it will require credentials and ask if you want to save them.}
\usage{
check_credentials()
}
\description{
Function to check is credentials are stored in a file.
It also provides the access token
If not, it will require credentials and ask if you want to save them.
}
\examples{
myaccesstoken=check_credentials()
myaccesstoken
}
