% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_device_list.r
\name{get_device_list}
\alias{get_device_list}
\title{Function to get a list of devices.}
\usage{
get_device_list()
}
\description{
Function to get a list of devices.
}
\examples{
devices=get_device_list()
devices
}
