% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home.R
\name{get_home}
\alias{get_home}
\title{Function to get the home directory as a function of the platform}
\usage{
get_home()
}
\description{
Function to get the home directory as a function of the platform
}
\examples{
get_home()
}
