% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myloadpackage.R
\name{myloadpackage}
\alias{myloadpackage}
\title{Function to install, if not already installed, and load a package}
\usage{
myloadpackage(x)
}
\description{
Function to install, if not already installed, and load a package
}
\examples{
myloadpackage(c('maps','ggplot2'))
}
