% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_kineis_data.r
\name{download_kineis_data}
\alias{download_kineis_data}
\title{Function to download data for a platform between a range of dates.}
\usage{
download_kineis_data(device_refs, from_date, to_date)
}
\description{
Function to download data for a platform between a range of dates.
}
\examples{
kineisdata=download_kineis_data(device_refs='267094',from_date="2025-10-15",to_date="2025-10-15")
head(kineisdata)
}
